+++ src/main/kotlin/com/github/gib/actions/GFavoriteWebAction.kt
@@ -2,13 +2,12 @@ package com.github.gib.actions
 
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.project.DumbAware
 import com.intellij.ui.jcef.JBCefBrowser
 import com.intellij.util.ui.JBImageIcon
 
 
 class GFavoriteWebAction(private val webUrl: String, icon: JBImageIcon, private val jbCefBrowser: JBCefBrowser) :
-    AnAction(icon), DumbAware {
+    AnAction(icon) {
 
     override fun actionPerformed(e: AnActionEvent) {
         jbCefBrowser.cefBrowser.loadURL(webUrl)