+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Add Zoom In & Out actions
 - Add Find Action
+- Initial work to support cookies manager in the browser
 
 ## [1.0.21]
 ### Added