+++ CHANGELOG.md
@@ -6,6 +6,14 @@
 - Add Find Action
 - Initial work to support cookies manager in the browser
 
+### Changed
+- Set Devtools enable by default
+- Toolwindow is now at left side by default
+
+### Fixed
+- Fix the issue with the devtools
+- Fix the issue with adding to favorites
+
 ## [1.0.21]
 ### Added
 - Support for Open in Browser