+++ CHANGELOG.md
@@ -1,7 +1,19 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Add Zoom In & Out actions
+- Add Find Action
+- Initial work to support cookies manager in the browser
 
+### Changed
+- Set Devtools enable by default
+- Toolwindow is now at left side by default
+
+### Fixed
+- Fix the issue with the devtools
+- Fix the issue with adding to favorites
+- Fix the change log job to be run as part of the release process
 ## [1.0.23]
 ### Added